Simplify github releaser

+++ sdks/github-java-sdk/src/main/java/org/jreleaser/sdk/github/GithubReleaser.java
@@ -17,7 +17,6 @@
  */
 package org.jreleaser.sdk.github;
 
-import org.jreleaser.model.Changelog;
 import org.jreleaser.model.JReleaserModel;
 import org.jreleaser.model.releaser.AbstractReleaserBuilder;
 import org.jreleaser.model.releaser.ReleaseException;
@@ -31,9 +30,6 @@ import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.List;
 
-import static org.jreleaser.util.StringUtils.isNotBlank;
-import static org.jreleaser.util.StringUtils.requireNonBlank;
-
 /**
  * @author Andres Almiray
  * @since 0.1.0
@@ -41,57 +37,36 @@ import static org.jreleaser.util.StringUtils.requireNonBlank;
 public class GithubReleaser implements Releaser {
     private final Logger logger;
     private final Path basedir;
-    private final String repo;
-    private final String commitsUrl;
-    private final String authorization;
-    private final String tagName;
-    private final String targetCommitish;
-    private final String releaseName;
-    private final Changelog changelog;
-    private final boolean draft;
-    private final boolean prerelease;
-    private final boolean overwrite;
-    private final boolean allowUploadToExisting;
-    private final String apiEndpoint;
+    private final JReleaserModel model;
     private final List<Path> assets = new ArrayList<>();
 
-    public GithubReleaser(Path basedir, Logger logger, String repo, String commitsUrl, String authorization,
-                          String tagName, String targetCommitish, String releaseName,
-                          Changelog changelog, boolean draft, boolean prerelease, boolean overwrite,
-                          boolean allowUploadToExisting, String apiEndpoint, List<Path> assets) {
+    public GithubReleaser(Path basedir, Logger logger, JReleaserModel model, List<Path> assets) {
         this.basedir = basedir;
         this.logger = logger;
-        this.repo = repo;
-        this.commitsUrl = commitsUrl;
-        this.authorization = authorization;
-        this.tagName = tagName;
-        this.targetCommitish = targetCommitish;
-        this.releaseName = releaseName;
-        this.changelog = changelog;
-        this.draft = draft;
-        this.prerelease = prerelease;
-        this.overwrite = overwrite;
-        this.allowUploadToExisting = allowUploadToExisting;
-        this.apiEndpoint = apiEndpoint;
+        this.model = model;
         this.assets.addAll(assets);
     }
 
-    public void release() throws ReleaseException {
-        Github api = new Github(logger, apiEndpoint, authorization);
+    public void release(boolean dryRun) throws ReleaseException {
+        org.jreleaser.model.Github github = model.getRelease().getGithub();
+
+        Github api = new Github(logger, github.getApiEndpoint(), github.getResolvedAuthorization());
+
+        String tagName = github.getTagName();
 
         try {
-            logger.info("Looking up release with tag {} at repository {}", tagName, repo);
-            GHRelease release = api.findReleaseByTag(repo, tagName);
+            logger.info("Looking up release with tag {} at repository {}", tagName, github.getCanonicalRepo());
+            GHRelease release = api.findReleaseByTag(github.getCanonicalRepo(), tagName);
             if (null != release) {
                 logger.info("Release {} exists", tagName);
-                if (overwrite) {
+                if (github.isOverwrite()) {
                     logger.info("Deleting release {}", tagName);
-                    release.delete();
+                    if (!dryRun) release.delete();
                     logger.info("Creating release {}", tagName);
-                    createRelease(api);
-                } else if (allowUploadToExisting) {
+                    createRelease(api, dryRun);
+                } else if (github.isAllowUploadToExisting()) {
                     logger.info("Updating release {}", tagName);
-                    api.uploadAssets(release, assets);
+                    if (!dryRun) api.uploadAssets(release, assets);
                 } else {
                     throw new IllegalStateException("Github release failed because release " +
                         tagName + " already exists. overwrite = false; allowUploadToExisting = false");
@@ -99,20 +74,26 @@ public class GithubReleaser implements Releaser {
             } else {
                 logger.info("Release {} does not exist", tagName);
                 logger.info("Creating release {}", tagName);
-                createRelease(api);
+                createRelease(api, dryRun);
             }
         } catch (IOException | IllegalStateException e) {
             throw new ReleaseException(e);
         }
     }
 
-    private void createRelease(Github api) throws IOException {
-        GHRelease release = api.createRelease(repo, tagName)
-            .commitish(targetCommitish)
-            .name(releaseName)
-            .draft(draft)
-            .prerelease(prerelease)
-            .body(ChangelogProvider.getChangelog(basedir, commitsUrl, changelog))
+    private void createRelease(Github api, boolean dryRun) throws IOException {
+        org.jreleaser.model.Github github = model.getRelease().getGithub();
+
+        String changelog = ChangelogProvider.getChangelog(basedir, github.getResolvedCommitUrl(), github.getChangelog());
+        logger.info("changelog:{}{}", System.lineSeparator(), changelog);
+        if (dryRun) return;
+
+        GHRelease release = api.createRelease(github.getCanonicalRepo(), github.getTagName())
+            .commitish(github.getTargetCommitish())
+            .name("Release " + model.getProject().getVersion())
+            .draft(github.isDraft())
+            .prerelease(github.isPrerelease())
+            .body(changelog)
             .create();
         api.uploadAssets(release, assets);
     }
@@ -124,54 +105,11 @@ public class GithubReleaser implements Releaser {
     }
 
     public static class Builder extends AbstractReleaserBuilder<GithubReleaser, Builder> {
-        private String targetCommitish = "main";
-        private boolean draft;
-        private boolean prerelease;
-        private String apiEndpoint = Github.ENDPOINT;
-
-        public Builder targetCommitish(String targetCommitish) {
-            this.targetCommitish = requireNonBlank(targetCommitish, "'targetCommitish' must not be blank");
-            return self();
-        }
-
-        public Builder draft(boolean draft) {
-            this.draft = draft;
-            return self();
-        }
-
-        public Builder prerelease(boolean prerelease) {
-            this.prerelease = prerelease;
-            return self();
-        }
-
-        public Builder apiEndpoint(String apiEndpoint) {
-            this.apiEndpoint = isNotBlank(apiEndpoint) ? apiEndpoint : Github.ENDPOINT;
-            return self();
-        }
-
         @Override
         public GithubReleaser build() {
             validate();
-            requireNonBlank(targetCommitish, "'targetCommitish' must not be blank");
 
-            return new GithubReleaser(basedir, logger, repo, commitsUrl, authorization,
-                tagName, targetCommitish, releaseName,
-                changelog, draft, prerelease, overwrite,
-                allowUploadToExisting, apiEndpoint, assets);
-        }
-
-        @Override
-        public Builder configureWith(Path basedir, JReleaserModel model) {
-            super.configureWith(basedir, model);
-
-            org.jreleaser.model.Github github = model.getRelease().getGithub();
-
-            targetCommitish(github.getTargetCommitish());
-            draft(github.isDraft());
-            prerelease(github.isPrerelease());
-            apiEndpoint(github.getApiEndpoint());
-
-            return self();
+            return new GithubReleaser(basedir, logger, model, assets);
         }
     }
 }
